Separate Creation of Main Layout into Website
